requirements.txt
# Auto-generated from pyproject.toml [tool.poetry.dependencies]
# This file contains only DIRECT dependencies, not transitive ones
# Source of truth: pyproject.toml
#
# To install:
#   pip install -r requirements.txt

backoff>=2.2.1,<3.0.0
beautifulsoup4>=4.12.3,<5.0.0
falkordb>=1.0.9,<2.0.0
fix-busted-json>=0.0.18,<0.0.19
litellm>=1.73.1,<2.0.0
pypdf>=5.9.0,<6.0.0
python-dotenv>=1.0.1,<2.0.0
ratelimit>=2.2.1,<3.0.0
requests>=2.32.3,<3.0.0
rich>=13.9.4,<14.0.0
typing-extensions>=4.12.1,<5.0.0
